Use integers for sceneRect calculations.
Previously we would often see a white strip a single pixel wide on
the right and/or bottom edge of the view. This was caused by QRectFs
being rounded to QRects for the actual painting to screen pixels. By
ensuring that the screneRect is composed of whole numbers, we eliminate
this issue.

svn path=/trunk/KDE/kdegames/kpat/; revision=1175046

+++ libkcardgame/kcardscene.cpp
@@ -432,13 +432,13 @@ void KCardScene::relayoutScene()
 
     d->deck->setCardWidth( n_scaleFactor * d->deck->cardWidth() );
 
-    qreal usedPixelWidth = usedWidth * d->deck->cardWidth();
-    qreal usedPixelHeight = usedHeight * d->deck->cardHeight();
-    qreal pixelHMargin = d->layoutMargin * d->deck->cardWidth();
-    qreal pixelVMargin = d->layoutMargin * d->deck->cardHeight();
+    int usedPixelWidth = usedWidth * d->deck->cardWidth();
+    int usedPixelHeight = usedHeight * d->deck->cardHeight();
+    int pixelHMargin = d->layoutMargin * d->deck->cardWidth();
+    int pixelVMargin = d->layoutMargin * d->deck->cardHeight();
 
-    qreal contentWidth;
-    qreal xOffset;
+    int contentWidth;
+    int xOffset;
     if ( d->alignment & AlignLeft )
     {
         contentWidth = usedPixelWidth;
@@ -460,8 +460,8 @@ void KCardScene::relayoutScene()
         xOffset = pixelHMargin;
     }
 
-    qreal contentHeight;
-    qreal yOffset;
+    int contentHeight;
+    int yOffset;
     if ( d->alignment & AlignTop )
     {
         contentHeight = usedPixelHeight;
